Allocate loading node from <lux/load.h> instead of <lux/htab.h>
The hash table implementation was not "symmetric" in the sense that it
allocates memory by using malloc(), but did not free memory.  Instead,
it returned a pointer to a node that need to be freed externally.  For
clarity, the new hash table implementation does not allocate memory.

Also, variables with specific types can use shorter names.