Start implement `lux-confgen` to automatically generate config codes
In C, it is common that configuration related codes are distributed in
multiple places in the source tree.  For example, there are header
files that define configuration structures, initialization functions
that fill in the default values, and configuration functions that fill
in the user-provided values based on, e.g., command line arguments or
configuration/setting files.  This is especially annoying in `lux`
because each of the `lux` module can in principle take a structure to
adjust its behavior at runtime.

`lux-confgen` is written to simplify this process.  The goal is that
we write a C-like parameter definition table, and `lux-confgen` will
automatically generate the configuration structure, initialization
function, and configuration function for us.  This can speed up
lux-module development and avoid bugs in configuration related codes.