Introduce <lux/algo.h> that mirrors <lux/task.h>
Important decision point: should task or algorithm be more fundamental
in lux?  If task is more fundamental, the planner should return tasks
(current design).  When running an algorithm is nothing but running a
task while adjusting parameters.  If algo is more fundamental (require
redesign), then task should be made just-in-time when we execute an
algorithm or submit a task.